WIP: data layer (Frame and Roll implementation, little changes to match related stuff) + upgrade script

+++ clients/android/bowling/src/main/java/fit/cvut/org/cz/bowling/data/entities/ParticipantStat.java
@@ -20,7 +20,7 @@ public class ParticipantStat extends fit.cvut.org.cz.tmlibrary.data.entities.Par
     private int score;
 
     @DatabaseField(columnName = DBConstants.cFRAMES_NUMBER)
-    private int framesPlayedNumber;
+    private byte framesPlayedNumber;
 
     public ParticipantStat() {}
